requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
dateparser==1.1.8
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
numpy==1.25.1
pandas==2.0.3
pip==22.3.1
pycryptodome==3.18.0
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
requests==2.31.0
setuptools==65.5.1
six==1.16.0
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.3
websockets==11.0.3
wheel==0.38.4
yarl==1.9.2
